+++ src/hive_mind_bridge/include/hive_mind_bridge/MessageUtils.h
