+++ src/hive_mind_bridge/src/main.cpp
@@ -1,5 +1,5 @@
-#include "hiveboard_bridge/HiveMindBridge.h"
-#include "hiveboard_bridge/MessageHandler.h"
+#include "hive_mind_bridge/HiveMindBridge.h"
+#include "hive_mind_bridge/MessageHandler.h"
 #include "ros/ros.h"
 #include "swarmus_ros_navigation/MoveByMessage.h"
 #include <hivemind-host/FunctionCallArgumentDTO.h>
@@ -9,7 +9,7 @@ constexpr uint8_t RATE_HZ{2};
 constexpr uint32_t compoundId{1}; // TODO find a way for the HiveMind and the robot to share this ID
 
 int main(int argc, char** argv) {
-    ros::init(argc, argv, "hiveboard_bridge");
+    ros::init(argc, argv, "hive_mind_bridge");
     ros::NodeHandle nodeHandle;
 
     std::string robotName = ros::param::param("~ROBOT_NAME", std::string("pioneer_0"));
